pre1

+++ src/build.sh
@@ -3,34 +3,42 @@ DEBUG=" -x "
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null && pwd )/../"
 set -euo pipefail
 
-[ -e "$DIR/src/covtotarget" ] && rm "$DIR/src/covtotarget"
-[ -e "$DIR/src/covtocounts" ] && rm "$DIR/src/covtocounts"
+for BINARY in covtotarget covtocounts covtobed2;
+do
+	[ -e "$DIR/src/$BINARY" ] && rm "$DIR/src/$BINARY"
+done
 
-echo " * Building test BAM"
+echo " * Building test BAM (from input/*.sam)"
 for i in $DIR/input/*.sam
 do
-  samtools view -bS $i | samtools sort -o ${i/sam/bam} -
-  samtools index  ${i/sam/bam}
+  if [ ! -e "${i/sam/bam}.bai" ]; then
+	  samtools view -bS $i | samtools sort -o ${i/sam/bam} -
+    samtools index  ${i/sam/bam}
+	fi
 done
+
 
 mkdir -p $DIR/{bin,release}/
-for SOURCE in covtocounts.nim covtotarget.nim;
+for SOURCE in covtocounts.nim covtotarget.nim covtobed2.nim;
 do
-	echo "[$SOURCE]"
+	echo "   $SOURCE"
 	OUTBIN=$(basename $SOURCE | cut -f1 -d.)
-	echo " * Compile (dyn. link)"
+	echo "   -  Compile dyn. link"
 	nim c -o:$DIR/bin/${OUTBIN}_debug $DIR/src/$SOURCE  > $DIR/.compile.log 2>&1 
 
-	echo " * Compile (static)"
-        cp $DIR/src/${SOURCE}ble $DIR/data.nimble
-	$DIR/bin/hts_nim_static_builder $DEBUG -s src/$SOURCE -n $DIR/data.nimble -- --opt:speed > $DIR/.docker.log 2>&1 && \
-	 sudo chown ubuntu $DIR/${OUTBIN} && \
-	 mv -v $DIR/${OUTBIN} $DIR/release/
-	rm $DIR/data.nimble
-        if [ ! -e $DIR/release/${OUTBIN} ]; then
-           echo "Static build failed."; exit;
-        fi
-
+	echo "   - Compile static (only ubuntu)"
+  if [ $USER == 'ubuntu' ]; then
+		cp $DIR/src/${SOURCE}ble $DIR/data.nimble
+		$DIR/bin/hts_nim_static_builder $DEBUG -s src/$SOURCE -n $DIR/data.nimble -- --opt:speed > $DIR/.docker.log 2>&1 && \
+		sudo chown ubuntu $DIR/${OUTBIN} && \
+		mv -v $DIR/${OUTBIN} $DIR/release/
+		rm $DIR/data.nimble
+					if [ ! -e $DIR/release/${OUTBIN} ]; then
+						echo "Static build failed."; exit;
+					fi
+  fi
 done
 
-sudo chown ubuntu:ubuntu $DIR/bin/*
+if [ $USER == 'ubuntu' ]; then
+  sudo chown ubuntu:ubuntu $DIR/bin/*
+fi
